+++ tests/test_270-system_policies.py
@@ -17,28 +17,30 @@ def test_create(cached_system_level_policy):
     assert cached_system_level_policy['name'].startswith('python-sdk')
 
 
-def test_create_single_nm():
+def test_create_single_nm(cached_tag):
     policy = britive.system.policies.build(
-        name=f'python-sdk',
+        name='python-sdk',
         tags=[cached_tag['name']],
         roles=['UserViewRole'],
         approval_notification_medium='Email',
         approver_tags=[cached_tag['name']]
     )
-    assert 'id' in policy.keys()
+
+    assert isinstance(policy, dict)
     assert 'name' in policy.keys()
     assert policy['name'].startswith('python-sdk')
 
 
-def test_create_multiple_nm():
+def test_create_multiple_nm(cached_tag):
     policy = britive.system.policies.build(
-        name=f'python-sdk',
+        name='python-sdk',
         tags=[cached_tag['name']],
         roles=['UserViewRole'],
         approval_notification_medium=['Email'],
         approver_tags=[cached_tag['name']]
     )
-    assert 'id' in policy.keys()
+
+    assert isinstance(policy, dict)
     assert 'name' in policy.keys()
     assert policy['name'].startswith('python-sdk')